Fixes to refresh() & configure() functions
The refresh() and configure() functions were coded to read the battery value (from the standard ZigBee cluster / attribute) and configure reporting on battery and temperature with intervals of 30 seconds minimum, 1 minute maximum. The code for configuring battery reporting had the incorrect dataType parameter and was causing a Java error.

I have reworked the refresh() function code to request an immediate reading of temperature and humidity, and also configure temp / humidity reporting with intervals of 30 minutes minimum, 6 hours maximum. Hopefully this will reduce the frequency of reporting and help conserve on battery power.

In the configure() function, I removed irrelevant comments, and replaced the previous readAttribute and configureReporting commands with a call to refresh(), to avoid unnecessarily duplicated/extraneous code. Also I added a debug log to inform that the DTH is configuring the device.