Fixes & Updates to preferences
I had a read through the SmartThings Developer Documentation on Preferences and Settings, and found that the required: true and displayDuringSetup: true are really meant for SmartApps, not device handlers. I tried using the Section () definition to separate different settings, but apparently that only works with SmartApps, not with device handlers.

Interestingly, on the Developer Documentation page for Device Handlers, there’s a note at the very bottom that recommends not using defaultValue for preferences. From all my testing, I disagree that it causes any issues, but conditional checks for null or "" values on preference setting variables should be left in the code just in case. This documentation web page also clearly shows that the preferences{} section should be INSIDE metadata{} and all examples show it comes after the definition() and tiles() sections.

Here are changes I've made:

In all DTHs:
• moved the preferences{} section inside metadata{}, below definition() and tiles()
• removed all section{} references
• removed all displayDuringSetup: true references
• changed all required: to false
• added default values to any preferences that didn’t have them
• reworked / edited the Date Format and Max/Volts description text for clarity and better formatting

In all DTHs with offset settings:
• added type: "decimal" (only allows input of numbers)
• range: "-X..Y" parameters, using appropriate X / Y values based on the type of offset

In the motion detector DTHs:
Motion Reset preference -
• added type: "number" (only allows input of integers without decimal values)
• added range: "1..7200"  to make sure only numerical values are input, but 0 and negative numbers are not allowed
• changed the default to 60 second to match the 60 second sensor hardware reset delay 
• edited the explanation text for clarity and moved it below the setting as a "NOTE:"

In the Temp/Humidity DTHs:
• added humidity offset preference - please double check parseHumidity() function for my changes!!